Corrigindo caregamento da lista de comentários
