+++ app/src/main/java/com/example/vitorpereira/projetofinal_android_ufg/CadastrarComentatio.java
@@ -3,10 +3,10 @@ package com.example.vitorpereira.projetofinal_android_ufg;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.net.Uri;
+import android.os.Bundle;
 import android.provider.MediaStore;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
