--- app/src/main/java/com/example/vitorpereira/projetofinal_android_ufg/DetalheReceitaActivity.java
@@ -1,13 +1,0 @@
-package com.example.vitorpereira.projetofinal_android_ufg;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-public class DetalheReceitaActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.fragment_activity_detalhe_receita);
-    }
-}
