+++ app/src/main/java/com/example/vitorpereira/projetofinal_android_ufg/LoginActivity.java
@@ -1,11 +1,10 @@
 package com.example.vitorpereira.projetofinal_android_ufg;
 
 import android.content.Intent;
+import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.widget.Toast;
-
 
 import com.firebase.ui.auth.AuthUI;
 import com.firebase.ui.auth.IdpResponse;
