+++ app/src/main/java/com/example/vitorpereira/projetofinal_android_ufg/ReceitaActivity.java
@@ -1,6 +1,10 @@
 package com.example.vitorpereira.projetofinal_android_ufg;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.Snackbar;
 import android.support.design.widget.TabLayout;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -14,10 +18,27 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+import android.widget.ListView;
 import android.widget.TextView;
+
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class ReceitaActivity extends AppCompatActivity {
 
+    private static DatabaseReference dataBaseReference;
+    private static StorageReference storageReference;
+
+    private static List<Comentario> listaComentarios = new ArrayList<Comentario>();
 
     private SectionsPagerAdapter mSectionsPagerAdapter;
     private ViewPager mViewPager;
@@ -43,6 +64,8 @@ public class ReceitaActivity extends AppCompatActivity {
         tabLayout.addOnTabSelectedListener(new TabLayout.ViewPagerOnTabSelectedListener(mViewPager));
 
         receita = (Receita) getIntent().getSerializableExtra("receita");
+
+        inicializeFirebase();
 
     }
 
@@ -79,8 +102,8 @@ public class ReceitaActivity extends AppCompatActivity {
 
         @Override
         public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-            View viewReceitaDetail = inflater.inflate(R.layout.fragment_activity_detalhe_receita, container, false);
-            View viewComment = inflater.inflate(R.layout.fragment_activity_comments, container, false);
+            final View viewReceitaDetail = inflater.inflate(R.layout.fragment_activity_detalhe_receita, container, false);
+            final View viewComment = inflater.inflate(R.layout.fragment_activity_comments, container, false);
 
             Receita receitaDetalhe = new Receita();
 
@@ -100,10 +123,50 @@ public class ReceitaActivity extends AppCompatActivity {
 
                 return viewReceitaDetail;
             } else if (getArguments().getInt(ARG_SECTION_NUMBER) == 2) {
+                FloatingActionButton fab = (FloatingActionButton) viewComment.findViewById(R.id.fab);
+                fab.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
+                        Intent intent = new Intent(viewComment.getContext(), CadastrarComentatio.class);
+                        startActivity(intent);
+                        Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
+                                .setAction("Action", null).show();
+                    }
+                });
+                listaDados(viewComment);
                 return viewComment;
             }
             return null;
         }
+
+    }
+
+    private void inicializeFirebase() {
+        FirebaseApp.initializeApp(this);
+        dataBaseReference = FirebaseDatabase.getInstance().getReference();
+        storageReference = FirebaseStorage.getInstance().getReference();
+    }
+
+    private static void listaDados(final View view) {
+        dataBaseReference.child("Comentarios").addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                listaComentarios.clear();
+                for (DataSnapshot dataSnapshot1 : dataSnapshot.getChildren()) {
+                    Comentario comentario = dataSnapshot1.getValue(Comentario.class);
+                    listaComentarios.add(comentario);
+                }
+
+                ListView lista = (ListView) view.findViewById(R.id.listview_coment_id);
+                ComentarioAdapter adapter = new ComentarioAdapter(view.getContext(), listaComentarios);
+                lista.setAdapter(adapter);
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+
+            }
+        });
     }
 
     public class SectionsPagerAdapter extends FragmentPagerAdapter {
