+++ app/src/main/java/com/example/vitorpereira/projetofinal_android_ufg/SplashScreenActivity.java
@@ -1,9 +1,9 @@
 package com.example.vitorpereira.projetofinal_android_ufg;
 
 import android.content.Intent;
+import android.os.Bundle;
 import android.os.Handler;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 
 
 public class SplashScreenActivity extends AppCompatActivity {
